sync with source

+++ truffle/src/com.oracle.truffle.api.interop/src/com/oracle/truffle/api/interop/DefaultIntegerExports.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2020, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -74,7 +74,7 @@ final class DefaultIntegerExports {
     static boolean fitsInFloat(Integer receiver) {
         int i = receiver;
         float f = i;
-        return f == i;
+        return (int) f == i;
     }
 
     @ExportMessage
@@ -101,7 +101,7 @@ final class DefaultIntegerExports {
     static float asFloat(Integer receiver) throws UnsupportedMessageException {
         int i = receiver;
         float f = i;
-        if (f == i) {
+        if ((int) f == i) {
             return f;
         }
         throw UnsupportedMessageException.create();
